Remove retries for slow predictions
These were added because predictions would sometimes get pre-empted and
then take 30 minutes to be reported as failed. Pre-empted predictions
should now be reported as failed within about 10 seconds.

Additionally, this retry behaviour is causing a huge queue to build up.
For some reason (as yet unknown) the deployment got a little bit behind,
to the point where it was actually taking longer than 5 minutes to
complete a prediction. Once it passed that point this logic repeatedly
requeued predictions that were already in the queue, waiting to be
processed.

If we need to reintroduce behaviour like this, we either need to send a
cancellation request or to only restart predictions that have been
*processing* for longer than 5 minutes (as opposed to starting).